Bind recyclerviewadapter with API response and probabs other things

+++ app/src/main/java/com/letsmeatapp/letsmeatapp/data/responses/RestaurantStatistics.kt
@@ -1,8 +1,8 @@
 package com.letsmeatapp.letsmeatapp.data.responses
 
 data class RestaurantStatistics(
-    val cleanness: Int,
-    val prices: Int,
-    val savouriness: Int,
-    val service: Int
+    val cleanness: Double,
+    val prices: Double,
+    val savouriness: Double,
+    val service: Double
 )
